Moved all angel.js code into a AngelCheckout.js #36
